# functions/docs/swagger.yaml
openapi: "3.0.0"
info:
  title: "전기안전 - Firebase Functions API"
  version: "1.0.0"
  description: |
    ## 🚀 Endpoint
    배포가 완료되면 아래 주소에서 API를 확인할 수 있습니다.

    - `https://us-central1-electrical-safety-4c9bd.cloudfunctions.net/api`
      <br>  <br>
    - `https://api-2hj4xnf47q-uc.a.run.app`
paths:
  /auth/admin/signup:
    post:
      tags:
        - Admin
      summary: "관리자 회원가입"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin_id
                - admin_pw
              properties:
                admin_id:
                  type: string
                  description: "관리자 ID"
                  example: "admin123"
                admin_pw:
                  type: string
                  description: "관리자 비밀번호"
                  example: "securePassword123!"
      responses:
        "201":
          description: "관리자 회원가입 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "관리자 회원가입 성공"
        "400":
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -2000
                  message:
                    type: string
                    example: "admin_id와 admin_pw를 입력하세요."
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -2002
                  message:
                    type: string
                    example: "관리자 회원가입 중 서버 오류 발생"
  /auth/admin/login:
    post:
      tags:
        - Admin
      summary: "관리자 로그인"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin_id
                - admin_pw
              properties:
                admin_id:
                  type: string
                  description: "관리자 ID"
                  example: "admin123"
                admin_pw:
                  type: string
                  description: "관리자 비밀번호"
                  example: "securePassword123!"
      responses:
        "200":
          description: "관리자 로그인 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "관리자 로그인 성공"
                  token:
                    type: string
                    example: "jwt_token_here"
                  lastActiveAt:
                    type: string
                    format: date-time
                    example: "2025-03-24T12:00:00.000Z"
        "400":
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -2000
                  message:
                    type: string
                    example: "admin_id와 admin_pw를 입력하세요."
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -2004
                  message:
                    type: string
                    example: "관리자 로그인 처리 중 서버 오류 발생"
  /auth/admin/{admin_id}:
    delete:
      tags:
        - Admin
      summary: "관리자 탈퇴"
      parameters:
        - name: admin_id
          in: path
          required: true
          schema:
            type: string
          description: "탈퇴할 관리자 ID"
      responses:
        "200":
          description: "관리자 탈퇴 성공"
        "400":
          description: "잘못된 요청"
        "500":
          description: "서버 오류"
  /auth/admins:
    get:
      tags:
        - Admin
      summary: "모든 관리자 목록 조회"
      responses:
        "200":
          description: "관리자 목록 반환"
        "404":
          description: "관리자가 없음"
        "500":
          description: "서버 오류"
  /auth/signup:
    post:
      tags:
        - User
      summary: "회원가입"
      description: "새로운 사용자를 등록하는 API입니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - user_pw
              properties:
                user_id:
                  type: string
                  description: "사용자의 고유 아이디"
                  example: "user123"
                user_pw:
                  type: string
                  description: "사용자의 비밀번호"
                  example: "password123"
      responses:
        201:
          description: "회원가입 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "회원가입 성공"
        400:
          description: "잘못된 요청 (user_id 또는 user_pw가 누락된 경우)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1000
                  message:
                    type: string
                    example: "user_id와 user_pw를 입력하세요."
        409:
          description: "이미 존재하는 사용자 ID"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1001
                  message:
                    type: string
                    example: "이미 존재하는 user_id입니다."
        500:
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1002
                  message:
                    type: string
                    example: "회원가입 중 서버 오류가 발생했습니다."
  /auth/login:
    post:
      tags:
        - User
      summary: "로그인"
      description: "user_id와 user_pw로 로그인하고 JWT 토큰을 발급받는 API입니다."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - user_pw
              properties:
                user_id:
                  type: string
                  description: "사용자의 고유 아이디"
                  example: "user123"
                user_pw:
                  type: string
                  description: "사용자의 비밀번호"
                  example: "password123"
      responses:
        200:
          description: "로그인 성공 (JWT 토큰 포함)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "로그인 성공"
                  token:
                    type: string
                    description: "JWT 토큰 (발급된 인증 토큰)"
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  lastActiveAt:
                    type: string
                    format: date-time
                    description: "마지막 로그인 시간"
                    example: "2025-01-23T12:34:56.789Z"
        400:
          description: "잘못된 요청 (user_id 또는 user_pw가 올바르지 않음)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1000
                  message:
                    type: string
                    example: "user_id와 user_pw를 입력하세요."
        404:
          description: "존재하지 않는 user_id"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1001
                  message:
                    type: string
                    example: "존재하지 않는 user_id입니다."
        401:
          description: "비밀번호가 일치하지 않음"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1003
                  message:
                    type: string
                    example: "비밀번호가 일치하지 않습니다."
        500:
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: -1004
                  message:
                    type: string
                    example: "로그인 처리 중 서버 오류가 발생했습니다."
  /auth/{user_id}:
    delete:
      tags:
        - User
      summary: "사용자 탈퇴"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: "탈퇴할 사용자 ID"
      responses:
        "200":
          description: "사용자 탈퇴 성공"
        "400":
          description: "잘못된 요청"
        "500":
          description: "서버 오류"
  /auth/users:
    get:
      tags:
        - User
      summary: "모든 사용자 목록 조회"
      responses:
        "200":
          description: "사용자 목록 반환"
        "404":
          description: "사용자가 없음"
        "500":
          description: "서버 오류"
...